+++ drevo/templates/drevo/knowledge_grade.html
@@ -17,8 +17,8 @@
         </div>
         <div class="col-2 ">
             <a style="margin-bottom:10px;" href="{% url 'grade_statistic' knowledge.id %}" class="btn btn-secondary">Статистика оценок</a>
-            <a href="{% url 'group_knowledge' knowledge.id %}" class="btn btn-secondary">Групповая оценка знания</a>
-            <a style="margin-top:10px;" href="{% url 'grade_infographics' knowledge.id %}" class="btn btn-secondary">Инфографика</a>
+            <a href="{% url 'group_knowledge' knowledge.id %}?variant={{ request.GET.variant }}" class="btn btn-secondary">Групповая оценка знания</a>
+            <a style="margin-top:10px;" href="{% url 'grade_infographics' knowledge.id %}?variant={{ request.GET.variant }}" class="btn btn-secondary">Инфографика</a>
         </div> 
     </div>
     <h2 class="my-2">{{ title }}</h2>
