+++ drevo/templates/drevo/knowledge_grade_statistic.html
@@ -15,6 +15,15 @@
             </div>
         </div>
     </a>
+    <div class="row justify-content-between">
+        <div class="col-auto ">
+            <a class="text-decoration-none" href="{{ knowledge.get_absolute_url }}/grade">
+                <i class="bi bi-arrow-left" style="font-size: 1.4rem; font-style: normal;">
+                        Оценка знания 
+                </i>           
+            </a>
+        </div>
+    </div>
     <h2 class="my-2">{{ title }}</h2>
 
     {% if not user.is_authenticated %}
