+++ drevo/urls.py
@@ -34,6 +34,7 @@ from .views import (
     GroupKnowledgeView,
     QuizDetailView,
     InfographicsView,
+    GroupInfographicsView,
     parameter_settings,
 )
 from .views import send_znanie, knowledge_feed_view
@@ -79,6 +80,7 @@ urlpatterns = [
     path("znanie/<int:pk>/grade/", KnowledgeFormView.as_view(), name="grade"),
     path('znanie/<int:pk>/grade/statistic', KnowledgeStatisticFormView.as_view(), name='grade_statistic'),
     path('znanie/<int:pk>/grade/group', GroupKnowledgeView.as_view(), name="group_knowledge"),
+    path('znanie/<int:pk>/grade/group/infographics', GroupInfographicsView.as_view(), name="grade_group_infographics"),
     path("znanie/<int:pk>/grade/infographics", InfographicsView.as_view(), name="grade_infographics"),
     path("all_quizzes/", QuizListView.as_view(), name="all_quizzes"),
     path("quiz/<int:pk>", QuizDetailView.as_view(), name="quiz"),
